+++ Learning English language.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.30907.101
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LElWPF.Core", "LElWPF.Core\LElWPF.Core.csproj", "{BCC6771A-A2F4-4EB3-B3D7-6081BE021B50}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{972E83FD-366E-4384-95AD-E1E7516C93EF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{972E83FD-366E-4384-95AD-E1E7516C93EF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "test2", "test2\test2.csproj", "{8A56623F-9524-47CE-9787-B36DD461200F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{972E83FD-366E-4384-95AD-E1E7516C93EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{972E83FD-366E-4384-95AD-E1E7516C93EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{972E83FD-366E-4384-95AD-E1E7516C93EF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8A56623F-9524-47CE-9787-B36DD461200F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8A56623F-9524-47CE-9787-B36DD461200F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8A56623F-9524-47CE-9787-B36DD461200F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8A56623F-9524-47CE-9787-B36DD461200F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
